+++ Wire-iOS/Sources/Authentication/Landing/LandingViewController.swift
@@ -10,7 +10,6 @@
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. dsSee the
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
